Use int instead of NSInteger

+++ src/objc/smtp/MCOSMTPSession.h
@@ -69,7 +69,7 @@
 
 @property (nonatomic, copy, readonly) NSString *lastSMTPResponse;
 
-@property (nonatomic, assign, readonly) NSInteger lastSMTPResponseCode;
+@property (nonatomic, assign, readonly) int lastSMTPResponseCode;
 
 /**
  Sets logger callback. The network traffic will be sent to this block.
